requirements.txt
asgiref==3.2.7
Babel==2.8.0
Django==3.0.6
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
phonenumbers==8.12.3
Pillow==7.1.2
pytz==2020.1
sqlparse==0.3.1
